Implement Update Book entity with author and category relationship.

+++ src/main/java/com/bookstore/onlinebookstore_service/entity/Book.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Entity
 @Getter
 @Setter
@@ -18,7 +20,7 @@ public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "book_id")
-    private Long bookId;
+    private Integer bookId;
     private String title;
 
     private Double price;
